server/app/requirements.txt
fastapi
uvicorn[standard]
httpx
sqlmodel
psycopg2-binary
python-dotenv
nltk